[HER-11]Follow-up Rearrange files
