+++ src/home/arrows/hooks/use-reset-state.js
@@ -1,4 +1,4 @@
-import { useContext } from 'home/carousel/hooks/use-context'
+import { useContext } from 'home/context'
 
 export const useResetState = () => {
   const { setSlideIn, setSlideDirection } = useContext()
