+++ src/home/arrows/hooks/use-set-timeout.js
@@ -1,4 +1,4 @@
-import { useContext } from 'home/carousel/hooks/use-context'
+import { useContext } from 'home/context'
 
 export const useSetTimeout = () => {
   const { setIndex, setSlideDirection, setSlideIn } = useContext()
