+++ src/home/content.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import { Title } from 'home/title'
+import { Carousel } from 'home/carousel'
+
+export const Content = () => <>
+  <Title />
+  <Carousel />
+</>
