+++ src/home/index.jsx
@@ -1,3 +1,8 @@
 import React from 'react'
 
-export const Home = () => <h1 />
+import { Context, useDefaultContext } from 'home/context'
+import { Content } from 'home/content'
+
+export const HomePage = () => <Context.Provider value={useDefaultContext()}>
+  <Content />
+</Context.Provider>
