+++ src/home/slider/style.js
