+++ src/home/title/index.jsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+import { useStyle } from 'home/title/style'
+
+export const Title = () => <h1 className={useStyle().head}>Hermes, we build the world</h1>
