+++ src/home/title/style.js
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core'
+
+const head = {
+  textAlign: 'center',
+  padding: '50px',
+  justifyContent: 'center',
+}
+
+export const useStyle = makeStyles({ head })
